Remove special cases from instrumentation code
Instead of having a special case for If-Return constructs, there are now
three locations: BEFORE, REPLACE, and AFTER. Instrumentation points for
If-Return now uses REPLACE to replace a line of code.
